+++ CHANGELOG.md
@@ -1,3 +1,6 @@
-## 1.0.0
+## 0.0.2
+* RouteTransition delimited to animated and none (linear, curve and fade removed). Support for curves added in the transition animation.
+* FxRouter: Basic navigations actions added (goTo, goToAndReplace, back, backUntil).
 
-* TODO: Describe initial release.
+## 0.0.1
+* Initial implementation of the package. Writing definitions, documentation and foundation methods.